pyproject.toml
[project]
name = "alx-backend-graphql-crm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-filter>=25.1",
    "graphene-django>=3.2.3",
]
